Refactor redux in to separate files
