+++ src/PortfolioPlanning/Redux/Contracts.ts
@@ -1,0 +1,7 @@
+import { IProject, IEpic } from "../Contracts";
+
+export interface IPortfolioPlanningState {
+    projects: IProject[];
+    epics: IEpic[];
+    message: string;
+}
