+++ src/PortfolioPlanning/Redux/Selectors/EpicTimelineSelectors.ts
@@ -1,0 +1,5 @@
+import { IPortfolioPlanningState } from "../Contracts";
+
+export function getMessage(state: IPortfolioPlanningState): string {
+    return state.message;
+}
